requirements.txt
elasticsearch>=8,<9
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
redshift-connector>=2.0.0
flask>=3.0.0
